requirements.txt
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.39.1
django-appconf==1.0.3
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-dirtyfields==1.3.1
django-forms-bootstrap==3.1.0
django-phone-verify==1.0.0
django-phonenumber-field==4.0.0
django-phonenumbers==1.0.1
django-plugins==0.3.0
django-sql==1.0.1
django-storages==1.6.5
djangorestframework==3.11.0
phonenumbers==8.11.2
psycopg2==2.8.4
pytz==2018.5
requests==2.21.0
sqlparse==0.2.4
stripe==2.27.0
django-heroku==0.3.1
gunicorn==19.7.1
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1

